refactor: move TemplatesDataSource to domain folder
