+++ app/api/core/application/propertyCreatorService/PropertyCreatorServiceStrategy.ts
@@ -1,5 +1,5 @@
 import { Context, Property, PropertyTypes } from 'api/templates.v2/model/Property';
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { RelationshipTypesDataSource } from 'api/relationshiptypes.v2/contracts/RelationshipTypesDataSource';
 import { SettingsDataSource } from 'api/settings.v2/contracts/SettingsDataSource';
 import { ArrayUtils } from 'api/common.v2/utils/Array';
