+++ app/api/entities.v2/database/ElasticEntityMapper.ts
@@ -1,4 +1,4 @@
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { RelationshipProperty } from 'api/templates.v2/model/RelationshipProperty';
 import { EntitySchema } from 'shared/types/entityType';
