+++ app/api/entities.v2/services/EntityRelationshipsUpdateService.ts
@@ -2,7 +2,7 @@ import { MatchQueryNode } from 'api/relationships.v2/model/MatchQueryNode';
 import { Template } from 'api/templates.v2/model/Template';
 import { RelationshipProperty } from 'api/templates.v2/model/RelationshipProperty';
 import { RelationshipsDataSource } from 'api/relationships.v2/contracts/RelationshipsDataSource';
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { Entity, EntityMetadata } from '../model/Entity';
 import { EntitiesDataSource } from '../contracts/EntitiesDataSource';
