+++ app/api/externalIntegrations.v2/automaticTranslation/GenerateAutomaticTranslationConfig.ts
@@ -1,4 +1,4 @@
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { ATConfigDataSource } from './contracts/ATConfigDataSource';
 import { GenerateATConfigError } from './errors/generateATErrors';
 import { Validator } from './infrastructure/Validator';
