+++ app/api/relationships.v2/services/GetRelationshipMigrationFieldService.ts
@@ -1,5 +1,5 @@
 import { TransactionManager } from 'api/common.v2/contracts/TransactionManager';
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { V1RelationshipProperty } from 'api/templates.v2/model/V1RelationshipProperty';
 import { objectIndex } from 'shared/data_utils/objectIndex';
 import { RelationshipMigrationFieldsDataSource } from '../contracts/RelationshipMigrationFieldsDataSource';
