+++ app/api/relationships.v2/services/GetRelationshipService.ts
@@ -3,7 +3,7 @@ import { EntitiesDataSource } from 'api/entities.v2/contracts/EntitiesDataSource
 import { Entity } from 'api/entities.v2/model/Entity';
 import { RelationshipTypesDataSource } from 'api/relationshiptypes.v2/contracts/RelationshipTypesDataSource';
 import { RelationshipType } from 'api/relationshiptypes.v2/model/RelationshipType';
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { Template } from 'api/templates.v2/model/Template';
 import { objectIndex } from 'shared/data_utils/objectIndex';
 import { RelationshipsDataSource } from '../contracts/RelationshipsDataSource';
