+++ app/api/relationships.v2/services/MigrationService.ts
@@ -3,7 +3,7 @@
 import { IdGenerator } from 'api/common.v2/contracts/IdGenerator';
 import { SaveStream } from 'api/common.v2/contracts/SaveStream';
 import { Logger } from 'api/log.v2/contracts/Logger';
-import { TemplatesDataSource } from 'api/templates.v2/contracts/TemplatesDataSource';
+import { TemplatesDataSource } from 'api/core/domain/template/TemplatesDataSource';
 import { objectIndexToArrays, objectIndexToSets } from 'shared/data_utils/objectIndex';
 import { TestOneHubRequest } from 'shared/types/api.v2/relationships.testOneHub';
 import { RelationshipsDataSource } from '../contracts/RelationshipsDataSource';
